+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v2.5.93
+
+### Fixed
+
+- `@icon` url in userscript's metadata [#301](https://github.com/AdguardTeam/PopupBlocker/issues/301)
+
 ## v2.5.92
 
 ### Fixed
 
-- conflicts with Wombat.js library [#291](https://github.com/AdguardTeam/PopupBlocker/issues/291).
+- conflicts with Wombat.js library [#291](https://github.com/AdguardTeam/PopupBlocker/issues/291)
 - github.com [#297](https://github.com/AdguardTeam/PopupBlocker/issues/297)
 - trailing spaces in meta headers [#301](https://github.com/AdguardTeam/PopupBlocker/issues/301)